Test resturcturing and temporary disabling.
