+++ test/field/common.py
@@ -1,0 +1,19 @@
+# encoding: utf-8
+
+from __future__ import unicode_literals
+
+import pytest
+
+from marrow.mongo import Document
+
+
+class FieldExam(object):
+	__args__ = tuple()
+	__kwargs__ = dict()
+	
+	@pytest.fixture()
+	def Sample(self, request):
+		class Sample(Document):
+			field = self.__field__(*self.__args__, **self.__kwargs__)
+		
+		return Sample
