+++ test/field/test_long.py
@@ -1,0 +1,14 @@
+# encoding: utf-8
+
+from __future__ import unicode_literals
+
+from common import FieldExam
+from marrow.mongo.field import Long
+
+
+class TestLongField(FieldExam):
+	__field__ = Long
+	
+	def test_biginteger(self, Sample):
+		result = Sample(272787482374844672646272463).field
+		assert result == 272787482374844672646272463
