+++ test/field/test_objectid.py
@@ -1,0 +1,55 @@
+# encoding: utf-8
+
+from __future__ import unicode_literals
+
+from datetime import timedelta
+
+from bson import ObjectId as oid
+
+from common import FieldExam
+from marrow.mongo import Document
+from marrow.mongo.field import ObjectId
+from marrow.mongo.util import utcnow
+from marrow.schema.compat import unicode
+
+
+class TestObjectIdField(FieldExam):
+	__field__ = ObjectId
+	
+	def test_id_default(self):
+		class Sample(Document):
+			id = ObjectId('_id')
+		
+		assert isinstance(Sample().id, oid)
+	
+	def test_cast_string(self, Sample):
+		inst = Sample('5832223f927cc6c1a10609f7')
+		
+		assert isinstance(inst.__data__['field'], oid)
+		assert unicode(inst.field) == '5832223f927cc6c1a10609f7'
+	
+	def test_cast_oid(self, Sample):
+		v = oid()
+		inst = Sample(v)
+		
+		assert inst.__data__['field'] is v
+	
+	def test_cast_datetime(self, Sample):
+		v = utcnow().replace(microsecond=0)
+		inst = Sample(v)
+		
+		assert isinstance(inst.__data__['field'], oid)
+		assert inst.field.generation_time == v
+	
+	def test_cast_timedelta(self, Sample):
+		v = -timedelta(days=7)
+		r = (utcnow() + v).replace(microsecond=0)
+		inst = Sample(v)
+		
+		assert isinstance(inst.__data__['field'], oid)
+		assert inst.field.generation_time == r
+	
+	def test_cast_document(self, Sample):
+		v = {'_id': oid()}
+		inst = Sample(v)
+		assert inst.field == v['_id']
